Force text input to uppercase if caps lock is detected on macOS
