+++ osu.Framework/Input/GameWindowTextInput.cs
@@ -17,28 +17,7 @@ namespace osu.Framework.Input
             this.window = window;
         }
 
-        private void window_KeyPress(object sender, osuTK.KeyPressEventArgs e)
-        {
-            // Drop any keypresses if the control, alt, or windows/command key are being held.
-            // This is a workaround for an issue on macOS where osuTK will fire KeyPress events even
-            // if modifier keys are held.  This can be reverted when it is fixed on osuTK's side.
-            if (RuntimeInfo.OS == RuntimeInfo.Platform.MacOsx)
-            {
-                var state = osuTK.Input.Keyboard.GetState();
-                if (state.IsKeyDown(osuTK.Input.Key.LControl)
-                    || state.IsKeyDown(osuTK.Input.Key.RControl)
-                    || state.IsKeyDown(osuTK.Input.Key.LAlt)
-                    || state.IsKeyDown(osuTK.Input.Key.RAlt)
-                    || state.IsKeyDown(osuTK.Input.Key.LWin)
-                    || state.IsKeyDown(osuTK.Input.Key.RWin))
-                    return;
-                // arbitrary choice here, but it caters for any non-printable keys on an A1243 Apple Keyboard
-                if (e.KeyChar > 63000)
-                    return;
-            }
-
-            pending += e.KeyChar;
-        }
+        protected virtual void HandleKeyPress(object sender, osuTK.KeyPressEventArgs e) => pending += e.KeyChar;
 
         public bool ImeActive => false;
 
@@ -56,12 +35,12 @@ namespace osu.Framework.Input
 
         public void Deactivate(object sender)
         {
-            window.KeyPress -= window_KeyPress;
+            window.KeyPress -= HandleKeyPress;
         }
 
         public void Activate(object sender)
         {
-            window.KeyPress += window_KeyPress;
+            window.KeyPress += HandleKeyPress;
         }
 
         private void imeCompose()
